% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{backup_all}
\alias{backup_all}
\title{Backup everything about given datasets}
\usage{
backup_all(
  dataset_ids,
  file_directory = here::here("data-raw"),
  overwrite = TRUE,
  metadata = TRUE,
  expression = TRUE,
  dif_exp_contrasts = TRUE,
  dif_exp_values = TRUE,
  platforms = TRUE
)
}
\arguments{
\item{dataset_ids}{ids of datasets to be backed up}

\item{file_directory}{path to save the raw output files. "data-raw" is default
for easy package creation}

\item{overwrite}{Boolean. Should existing files be overwritten}

\item{metadata}{if TRUE, sample metadata will be backed up}

\item{expression}{if TRUE, expression data will be backed up}

\item{dif_exp_contrasts}{if TRUE, differential expression contrasts will be backed up}

\item{dif_exp_values}{if TRUE, sample differential expression values will be backed up}

\item{platforms}{if TRUE, platforms used with the datasets will be backed up}
}
\description{
This function backs up everything about listed datasets as raw files
}
