% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backups.R
\name{backup_datasets}
\alias{backup_datasets}
\title{Backup datasets}
\usage{
backup_datasets(
  dataset_ids,
  file_directory = here::here("data-raw"),
  overwrite = TRUE
)
}
\arguments{
\item{dataset_ids}{ids of datasets to be backed up}

\item{file_directory}{path to save the raw output files. "data-raw" is default
for easy package creation}

\item{overwrite}{Boolean. Should existing files be overwritten}
}
\description{
Saves raw outputs from the gemma API to be processed
}
