% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backups.R
\name{backup_differential_expression_contrasts}
\alias{backup_differential_expression_contrasts}
\title{Backup differential expression contrasts}
\usage{
backup_differential_expression_contrasts(
  dataset_ids,
  json_directory = here::here("data-raw"),
  overwrite = TRUE
)
}
\arguments{
\item{dataset_ids}{ids of datasets to be backed up}

\item{json_directory}{path to save the raw json files. "data-raw" is default
for easy package creation}

\item{overwrite}{Boolean. Should existing files be overwritten}
}
\description{
Saves raw outputs from the gemma API to be processed
}
