% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backups.R
\name{backup_plaftorms}
\alias{backup_plaftorms}
\title{Backup Platforms}
\usage{
backup_plaftorms(
  platform_ids,
  json_directory = here::here("data-raw"),
  overwrite = TRUE
)
}
\arguments{
\item{platform_ids}{ids of platforms to be backed up}

\item{json_directory}{path to save the raw json files. "data-raw" is default
for easy package creation}

\item{overwrite}{Boolean. Should existing files be overwritten}
}
\description{
Saves raw outputs from the gemma API to be processed
}
