% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backups.R
\name{backup_platform_annotations}
\alias{backup_platform_annotations}
\title{Backup platform annotations}
\usage{
backup_platform_annotations(
  platform_ids,
  file_directory = here::here("data-raw"),
  overwrite = TRUE
)
}
\arguments{
\item{platform_ids}{ids of platforms to be backed up}

\item{file_directory}{path to save the raw output files. "data-raw" is default
for easy package creation}

\item{overwrite}{Boolean. Should existing files be overwritten}
}
\description{
Saves raw outputs from the gemma API to be processed
}
