% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_data_load.R
\name{load_values}
\alias{load_values}
\title{Load expression and differential expression data}
\usage{
load_values(
  load = c("expression", "differential_expression"),
  envir = parent.frame()
)
}
\arguments{
\item{load}{Which datasets to load.}
}
\description{
Due to Github single file size limitations, expression and differential
expression data for the datasets are not automatically loaded in package
load. Use this function to load \code{\link{expression}} and \code{\link{differential_expression_values}}
objects
}
