% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backups.R
\name{package_rdas}
\alias{package_rdas}
\title{Populate "data" directory with backed up data}
\usage{
package_rdas(file_directory = here::here("data-raw"), overwrite = TRUE)
}
\description{
Populate "data" directory with backed up data
}
