% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R, R/z_processors.R
\name{processExpressionMatrix}
\alias{processExpressionMatrix}
\title{Processes expression matrix}
\usage{
processExpressionMatrix(m, samples)

processExpressionMatrix(m, samples)
}
\arguments{
\item{m}{The expression matrix to process}
}
\value{
A processed matrix

A processed matrix
}
\description{
Processes expression matrix

Processes expression matrix
}
\keyword{internal}
