% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processors.R, R/z_processors.R
\name{processFile}
\alias{processFile}
\title{Processes a response as a gzip file}
\usage{
processFile(content, ...)

processFile(content, ...)
}
\arguments{
\item{content}{The content from an \code{http_get} request}
}
\value{
A processed data.table

A processed data.table
}
\description{
Processes a response as a gzip file

Processes a response as a gzip file
}
\keyword{internal}
