% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backups.R
\name{backup_args}
\alias{backup_args}
\title{backup_args}
\arguments{
\item{platform_ids}{ids of platforms to be backed up}

\item{dataset_ids}{ids of datasets to be backed up}

\item{json_directory}{path to save the raw json files. "data-raw" is default
for easy package creation}

\item{overwrite}{Boolean. Should existing files be overwritten}
}
\description{
backup_args
}
